refactor: add infra package for test
